loans2grow/requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.1
phonenumbers==8.13.27
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
